+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `join`: Combine strings.
 - [#1292](http://github.com/nelson-lang/nelson/issues/1292) Large Table Display.
 - [#1290](http://github.com/nelson-lang/nelson/issues/1290) `VariableTypes` property for table: Specify the data types of table in Nelson.
+- `hour`, `minute`, `second` component of input date and time.
 
 ### Changed